requirements.txt
pandas
psycopg2
sqlalchemy
streamlit
python-dotenv
ipykernel